requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyYAML==6.0.2
pydantic==2.9.2
python-multipart==0.0.9
Pillow>=10.0.0
boto3>=1.28.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# UI
streamlit>=1.30.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# AWS
aioboto3>=12.0.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Monitoring
prometheus-client>=0.19.0
